django_project/requirements.txt
asgiref~=3.7.2
Django~=5.0.3
djangorestframework~=3.14.0
netifaces~=0.11.0
psycopg2~=2.9.9
python-dotenv~=1.0.1
sqlparse~=0.4.4
tzdata~=2024.1